requirements.txt
openai>=1.0.0
pyTelegramBotAPI
python-dotenv
pydub
requests